# openapi.yaml
openapi: 3.0.3
info:
  title: Armada Convoy API
  description: REST API for the Armada convoy and social app
  version: 1.0.0
  contact:
    name: Armada Team
    email: support@armada.app
servers:
  - url: https://convoy-api.onrender.com
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /convoys:
    get:
      summary: Get convoys
      description: Get list of convoys
      parameters:
        - name: limit
          in: query
          description: Number of convoys to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of convoys to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: lat
          in: query
          description: Latitude for location-based search
          schema:
            type: number
            format: float
        - name: lng
          in: query
          description: Longitude for location-based search
          schema:
            type: number
            format: float
        - name: radius
          in: query
          description: Search radius in kilometers
          schema:
            type: number
            default: 50
      responses:
        '200':
          description: Convoys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      convoys:
                        type: array
                        items:
                          $ref: '#/components/schemas/Convoy'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      summary: Create convoy
      description: Create a new convoy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentCenter
              properties:
                title:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                visibility:
                  type: string
                  enum: [public, invite, private]
                  default: public
                maxMembers:
                  type: integer
                  minimum: 2
                  maximum: 50
                  default: 20
                route:
                  $ref: '#/components/schemas/ConvoyRoute'
                currentCenter:
                  $ref: '#/components/schemas/ConvoyLocation'
      responses:
        '201':
          description: Convoy created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Convoy'
  /convoys/{id}/join:
    post:
      summary: Join convoy
      description: Join an existing convoy
      parameters:
        - name: id
          in: path
          required: true
          description: Convoy ID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                joinCode:
                  type: string
                  description: Join code for invite-only convoys
      responses:
        '200':
          description: Joined convoy successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Successfully joined convoy
  /convoys/{id}/location:
    post:
      summary: Update convoy location
      description: Update the convoy's current location
      parameters:
        - name: id
          in: path
          required: true
          description: Convoy ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvoyLocation'
      responses:
        '200':
          description: Location updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Location updated successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        avatarUrl:
          type: string
          format: uri
          example: https://example.com/avatar.jpg
        bio:
          type: string
          example: Adventure enthusiast and convoy leader
        settings:
          $ref: '#/components/schemas/UserSettings'
        stats:
          $ref: '#/components/schemas/UserStats'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSettings:
      type: object
      properties:
        showConvoys:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showStats:
          type: string
          enum: [everyone, friends, private]
          example: friends
        showProfile:
          type: string
          enum: [everyone, friends, private]
          example: everyone
        notifications:
          type: object
          properties:
            convoyInvites:
              type: boolean
              example: true
            friendRequests:
              type: boolean
              example: true
            messages:
              type: boolean
              example: true
            achievements:
              type: boolean
              example: true
    UserStats:
      type: object
      properties:
        convoysCreated:
          type: integer
          example: 12
        convoysJoined:
          type: integer
          example: 45
        totalMiles:
          type: number
          example: 1250.5
        friendsCount:
          type: integer
          example: 23
...
    Convoy:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        ownerId:
          type: string
          example: 507f1f77bcf86cd799439011
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        isLive:
          type: boolean
          example: true
        route:
          $ref: '#/components/schemas/ConvoyRoute'
        currentCenter:
          $ref: '#/components/schemas/ConvoyLocation'
        visibility:
          type: string
          enum: [public, invite, private]
          example: public
        joinCode:
          type: string
          example: ABC123
        title:
          type: string
          example: Weekend Road Trip
        description:
          type: string
          example: A scenic drive through the mountains
        maxMembers:
          type: integer
          example: 20
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConvoyRoute:
      type: object
      properties:
        waypoints:
          type: array
          items:
            $ref: '#/components/schemas/ConvoyWaypoint'
        polyline:
          type: string
          example: encoded_polyline_string
        distance:
          type: number
          example: 150.5
        duration:
          type: number
          example: 7200
    ConvoyWaypoint:
      type: object
      properties:
        lat:
          type: number
          format: float
          example: 37.7749
        lng:
          type: number
          format: float
          example: -122.4194
        name:
          type: string
          example: Golden Gate Bridge
        order:
          type: integer
          example: 1
    ConvoyLocation:
      type: object
      properties:
        lat:
          type: number
          format: float
          example: 37.7749
        lng:
          type: number
          format: float
          example: -122.4194
        heading:
          type: number
          format: float
          example: 45.0
        speed:
          type: number
          format: float
          example: 65.5
        accuracy:
          type: number
          format: float
          example: 10.0
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        total:
          type: integer
          example: 100
        hasMore:
          type: boolean
          example: true
...